requirements.txt
fastapi=0.110.1
uvicorn=0.29.0
motor=3.4.0
pydantic-settings=2.2.1